# api/openapi.yaml
openapi: 3.0.3
info:
  title: DAB Music Player API
  description: |
    A comprehensive music streaming and library management API that provides access to music search,
    album information, user authentication, playlist management, and administrative features.

    THIS SOFTWARE AND API ARE PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
    EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF
    MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT.
    IN NO EVENT SHALL THE AUTHORS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
    LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING
    FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER
    DEALINGS IN THE SOFTWARE OR API.

    By using this API or building on it, you agree that any software or service that incorporates
    it must also be made available under an open-source license compatible with AGPLv3.

    ## Authentication
    Most endpoints require authentication via JWT tokens stored in HTTP-only cookies.

    ## Base URL
    All API endpoints are relative to: `https://dab.yeet.su/api`
  version: 1.0.0
  contact:
    name: DAB Music Player
    url: https://dab.yeet.su
  license:
    name: GNU Affero General Public License v3.0
    url: https://www.gnu.org/licenses/agpl-3.0.en.html
servers:
  - url: https://dab.yeet.su/api
    description: Production server
security:
  - cookieAuth: []
  - {}
paths:
...
  /auth/login:
    post:
      tags:
        - Authentication
      summary: User login
      description: Authenticate user with email and password
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                  example: user@example.com
                password:
                  type: string
                  format: password
                  example: password123
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Login successful
                  user:
                    $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /auth/register:
    post:
      tags:
        - Authentication
      summary: User registration
      description: Register a new user account
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - username
                - email
                - password
              properties:
                username:
                  type: string
                  example: johndoe
                email:
                  type: string
                  format: email
                  example: john@example.com
                password:
                  type: string
                  format: password
                  example: password123
                inviteCode:
                  type: string
                  description: Required if invite system is enabled
                  example: INVITE123
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: User created successfully
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /auth/logout:
    post:
      tags:
        - Authentication
      summary: User logout
      description: Logout current user and clear session
      responses:
        '200':
          description: Logged out successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Logged out successfully
        '500':
          $ref: '#/components/responses/InternalServerError'
  /auth/me:
    get:
      tags:
        - Authentication
      summary: Get current user
      description: Get information about the currently authenticated user
      responses:
        '200':
          description: User information retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    allOf:
                      - $ref: '#/components/schemas/User'
                    nullable: true
        '500':
          $ref: '#/components/responses/InternalServerError'
  /auth/forgot-password:
    post:
      tags:
        - Authentication
      summary: Request password reset
      description: Send password reset email to user
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
              properties:
                email:
                  type: string
                  format: email
                  example: user@example.com
      responses:
        '200':
          description: Password reset email sent
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /auth/reset-password:
    post:
      tags:
        - Authentication
      summary: Reset password
      description: Reset user password using reset token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - token
                - password
              properties:
                token:
                  type: string
                  example: reset-token-123
                password:
                  type: string
                  format: password
                  example: newpassword123
      responses:
        '200':
          description: Password reset successful
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
  # Music Endpoints
  /search:
    get:
      tags:
        - Music
      summary: Search for music
      description: Search for tracks, albums, or artists
      security: []
      parameters:
        - name: q
          in: query
          required: true
          description: Search query
          schema:
            type: string
            example: "The Beatles"
        - name: type
          in: query
          description: Type of content to search for
          schema:
            type: string
            enum: [track, album, artist]
            default: track
        - name: limit
          in: query
          description: Number of results to return
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 20
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                type: object
                properties:
                  results:
                    type: array
                    items:
                      oneOf:
                        - $ref: '#/components/schemas/Track'
                        - $ref: '#/components/schemas/Album'
                        - $ref: '#/components/schemas/Artist'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /album:
    get:
      tags:
        - Music
      summary: Get album information
      description: Retrieve detailed information about an album
      security: []
      parameters:
        - name: albumId
          in: query
          required: true
          description: Album ID
          schema:
            type: string
            example: "12345"
      responses:
        '200':
          description: Album information
          content:
            application/json:
              schema:
                type: object
                properties:
                  album:
                    $ref: '#/components/schemas/Album'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /album/{id}:
    get:
      tags:
        - Music
      summary: Get album by ID
      description: Get album details from database
      parameters:
        - name: id
          in: path
          required: true
          description: Album ID
          schema:
            type: string
      responses:
        '200':
          description: Album found
          content:
            application/json:
              schema:
                type: object
                properties:
                  album:
                    $ref: '#/components/schemas/Album'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /discography:
    get:
      tags:
        - Music
      summary: Get artist discography
      description: Retrieve all albums by an artist
      security: []
      parameters:
        - name: artistId
          in: query
          required: true
          description: Artist ID
          schema:
            type: string
            example: "67890"
      responses:
        '200':
          description: Artist discography
          content:
            application/json:
              schema:
                type: object
                properties:
                  artist:
                    $ref: '#/components/schemas/Artist'
                  albums:
                    type: array
                    items:
                      $ref: '#/components/schemas/Album'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /download:
    get:
      tags:
        - Music
      summary: Get album download information
      description: Retrieve album information for download
      security: []
      parameters:
        - name: albumId
          in: query
          required: true
          description: Album ID
          schema:
            type: string
        - name: quality
          in: query
          description: Audio quality
          schema:
            type: string
            default: "27"
            example: "27"
      responses:
        '200':
          description: Album download information
          content:
            application/json:
              schema:
                type: object
                properties:
                  album:
                    $ref: '#/components/schemas/Album'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /stream:
    get:
      tags:
        - Music
      summary: Stream audio track
      description: Get streaming URL for a track
      security: []
      parameters:
        - name: trackId
          in: query
          required: true
          description: Track ID
          schema:
            type: string
        - name: quality
          in: query
          description: Audio quality
          schema:
            type: string
            default: "27"
      responses:
        '200':
          description: Stream URL
          content:
            application/json:
              schema:
                type: object
                properties:
                  streamUrl:
                    type: string
                    format: uri
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /lyrics:
    get:
      tags:
        - Music
      summary: Get song lyrics
      description: Retrieve lyrics for a song
      security: []
      parameters:
        - name: artist
          in: query
          required: true
          description: Artist name
          schema:
            type: string
            example: "The Beatles"
        - name: title
          in: query
          required: true
          description: Song title
          schema:
            type: string
            example: "Hey Jude"
      responses:
        '200':
          description: Lyrics found
          content:
            application/json:
              schema:
                type: object
                properties:
                  lyrics:
                    type: string
                    example: "Hey Jude, don't make it bad..."
                  unsynced:
                    type: boolean
                    description: Whether lyrics are time-synced
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  # User Features
...
  /libraries/{id}/tracks/{trackId}:
    delete:
      tags:
        - Libraries
      summary: Remove track from library
      description: Remove a track from a library
      parameters:
        - name: id
          in: path
          required: true
          description: Library ID
          schema:
            type: string
        - name: trackId
          in: path
          required: true
          description: Track ID
          schema:
            type: string
      responses:
        '200':
          description: Track removed from library
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: session
      description: JWT token stored in HTTP-only cookie
  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
          example: 1
        username:
          type: string
          example: johndoe
        email:
          type: string
          format: email
          example: john@example.com
        created_at:
          type: string
          format: date-time
...